Fix odd layout jumping when incrementing slider key
When incrementing the slider key for the focal point picker, a strange
jump would occur between initial mount and second mount. The jump
appeared to be related to rendering the "value component" in
`BottomSheet.Cell`. Replacing the empty `value` with the `leftAlign`
label fixes the issue and appears to create no regression.

- https://git.io/JTmWU