Dismiss bottom sheet before media picker displayed
This causes the media picker to display and then disappear. This likely
occurs because the picker's presence is contained within the bottom
sheet itself. Because of this relationship, this approach will likely
not work.

This exploration was done to support the new add/edit media buttons
found within the Cover block inspector controls bottom sheet.
`react-native-modal` does not support displaying multiple modals on iOS,
therefore we must identify a proper method of transitioning from one
modal to the next.