Dismiss bottom sheet after media picker has been displayed
This requires `MediaPicker` to be wrapped in the
`BottomSheetContextProvider`, so that the `MediaUpload` may register an
`onDismiss` callback.

This exploration was done to support the new add/edit media buttons
found within the Cover block inspector controls bottom sheet.
`react-native-modal` does not support displaying multiple modals on iOS,
therefore we must identify a proper method of transitioning from one
modal to the next.